+++ Assets/Scripts/DI/GameLifetimeScope.cs
@@ -1,5 +1,7 @@
 ﻿namespace Scripts.DI
 {
+    using Cuvara.Netcode.DI;
+    using Scripts.Nakama.DI;
     using VContainer;
     using VContainer.Unity;
 
@@ -8,6 +10,8 @@
         protected override void Configure(IContainerBuilder builder)
         {
             base.Configure(builder);
+            builder.RegisterNetworking();
+            builder.RegisterNakama();
         }
     }
 }